Readonly mode for forms; Roles and Permissions; Code refactoring

+++ src/Http/Controllers/ModuleController.php
@@ -44,7 +44,15 @@ class ModuleController extends BaseController
 
     public $description = '';
 
-    public $buttons = [];
+    public $actions = [
+        'add' => true,
+        'view' => false,
+        'edit' => true,
+        'delete' => true,
+        'multiple-delete' => true,
+    ];
+
+    public $readonly = false;
 
     public function before()
     {
@@ -69,6 +77,38 @@ class ModuleController extends BaseController
         }
     }
 
+    public function can(string $action)
+    {
+        return isset($this->actions[$action]) ? $this->actions[$action] : false;
+    }
+
+    public function isReadOnly()
+    {
+        return $this->readonly;
+    }
+
+    protected function getGridOptions()
+    {
+        $options = $this->grid;
+
+        if (!isset($options['add']))
+            $options['add'] = $this->can('add');
+
+        if (!isset($options['view']))
+            $options['view'] = $this->can('view');
+
+        if (!isset($options['edit']))
+            $options['edit'] = $this->can('edit');
+
+        if (!isset($options['delete']))
+            $options['delete'] = $this->can('delete');
+
+        if (!isset($options['multiple-delete']))
+            $options['multiple-delete'] = $this->can('multiple-delete');
+
+        return $options;
+    }
+
     public function index()
     {
         // sorting save
@@ -85,11 +125,11 @@ class ModuleController extends BaseController
                         $currentSort = [];
                         $items = array_flip($items);
 
-                        foreach ($list AS $item) {
+                        foreach ($list as $item) {
                             $currentSort[] = $item->sort_order;
                         }
 
-                        foreach ($list AS $item) {
+                        foreach ($list as $item) {
                             $newSortIndex = $currentSort[$items[$item->id]];
                             if ($newSortIndex) {
                                 $item->sort_order = $newSortIndex;
@@ -109,7 +149,7 @@ class ModuleController extends BaseController
             redirect()->to(request()->fullUrl())->send();
         }
 
-        $grid = new Grid($this->grid, $this->getSection());
+        $grid = new Grid($this->getGridOptions(), $this->getSection());
 
         if ($this->mode == 'simple')
             return view('cms::simple.module')
@@ -141,16 +181,25 @@ class ModuleController extends BaseController
         return $this->description ?: $default;
     }
 
+    protected function getModel(): Model
+    {
+        if (is_null($this->model) && $this->objectId)
+            $this->model = $this->className::findOrFail($this->objectId);
+
+        return $this->model;
+    }
+
     /**
      * Return form Builder object
      * @param boolean $create
      * @return Builder
      */
-    protected function getForm($create = false): Builder
+    protected function getForm(bool $create = true): Builder
     {
         $repository = null;
 
         $formFields = $this->formFields();
+
         if ($this->objectId) {
             $model = $this->className::findOrFail($this->objectId);
             $values = $model->getAttributes();
@@ -178,7 +227,7 @@ class ModuleController extends BaseController
                     $group = str_replace('[]', '', $field->get('name'));
 
                     if ($files) {
-                        $fieldFiles = $files->filter(function($item) use ($group) {
+                        $fieldFiles = $files->filter(function ($item) use ($group) {
                             return $item->group == $group;
                         });
 
@@ -192,55 +241,40 @@ class ModuleController extends BaseController
         }
 
         $form = new Builder($formFields, $repository);
+        if ($this->readonly)
+            $form->readonly(true);
         $form->groups($this->formGroups());
 
-        if ($create) {
-            $form->push(
-                Button::make('create')
-                    ->label('Добавить')
-                    ->value('true')
-                    ->class('btn btn-primary')
-            );
-            $form->setAction(route('cms.module.store', ['controller' => $this->getSectionController()], false));
+        if (!$this->readonly) {
+            if ($create) {
+                $form->push(
+                    Button::make('create')
+                        ->label('Добавить')
+                        ->value('true')
+                        ->class('btn btn-primary')
+                );
 
-        } else {
-            $form->push(
-                Button::make('update')
-                    ->label('Сохранить')
-                    ->value('true')
-                    ->class('btn btn-primary')
-            );
+            } else {
+                $form->push(
+                    Button::make('update')
+                        ->label('Сохранить')
+                        ->value('true')
+                        ->class('btn btn-primary')
+                );
 
-            if ($this->buttons) {
-                foreach ($this->buttons as $btn) {
-                    $form->push(
-                        Link::make($btn['title'])
-                            ->href($btn['link'])
-                            ->class($btn['class'])
-                            ->withoutFormType()
-                    );
-                }
+                $form->push(
+                    Link::make('Удалить')
+                        ->href(route('cms.module.destroy', ['controller' => $this->getSectionController(), 'objectId' => $this->objectId], false))
+                        ->class('btn btn-danger float-right')
+                        ->confirm('true')
+                        ->withoutFormType()
+                );
+
+                $form->images(route('cms.module.images', ['controller' => $this->getSectionController(), 'objectId' => $this->objectId], false));
             }
-
-            $form->push(
-                Link::make('Удалить')
-                    ->href(route('cms.module.destroy', ['controller' => $this->getSectionController(), 'objectId' => $this->objectId], false))
-                    ->class('btn btn-danger float-right')
-                    ->confirm('true')
-                    ->withoutFormType()
-            );
-
-            $form->images(route('cms.module.images', ['controller' => $this->getSectionController(), 'objectId' => $this->objectId], false));
-
-            $form->setAction(route('cms.module.update', [
-                'controller' => $this->getSectionController(),
-                'objectId' => $this->objectId
-            ], false));
-
-            $form->method('PUT');
         }
 
-        return $form->view('card');
+        return $form;
     }
 
     /**
@@ -250,14 +284,16 @@ class ModuleController extends BaseController
     {
         $form = $this->getForm(true);
 
+        $form->setAction(route('cms.module.store', ['controller' => $this->getSectionController()], false));
+
         if ($this->mode == 'simple')
             return view('cms::simple.module')
-                ->with('form', $form)
+                ->with('form', $form->view('card'))
                 ->with('title', $this->getTitle('Добавление'))
                 ->with('description', $this->getDescription());
         else
             return view('cms::module')
-                ->with('form', $form)
+                ->with('form', $form->view('card'))
                 ->with('title', $this->getTitle('Добавление'))
                 ->with('description', $this->getDescription());
     }
@@ -275,22 +311,37 @@ class ModuleController extends BaseController
     /**
      * Show the form for editing the specified resource.
      *
-     * @param  int $id
+     * @param int $id
      */
     public function edit($objectId)
     {
         $this->objectId = $objectId;
+        if ($this->getModel()->is_approved)
+            return redirect(
+                route(
+                    'cms.module.view',
+                    ['controller' => $this->getSectionController(), 'objectId' => $this->model->getKey()],
+                    false
+                ) . ($this->mode == 'simple' ? '?modal_parent_id=' . request()->get('modal_parent_id') : '')
+            );
 
-        $form = $this->getForm(false);
+        $form = $this->getForm();
+
+        $form->setAction(route('cms.module.update', [
+            'controller' => $this->getSectionController(),
+            'objectId' => $this->objectId
+        ], false));
+
+        $form->method('PUT');
 
         if ($this->mode == 'simple')
             return view('cms::simple.module')
-                ->with('form', $form)
+                ->with('form', $form->view('card'))
                 ->with('title', $this->getTitle('Редактирование'))
                 ->with('description', $this->getDescription());
         else
             return view('cms::module')
-                ->with('form', $form)
+                ->with('form', $form->view('card'))
                 ->with('title', $this->getTitle('Редактирование'))
                 ->with('description', $this->getDescription());
     }
@@ -304,6 +355,25 @@ class ModuleController extends BaseController
         }
 
         return false;
+    }
+
+    public function view($objectId)
+    {
+        $this->objectId = $objectId;
+        $this->readonly = true;
+
+        $form = $this->getForm();
+
+        if ($this->mode == 'simple')
+            return view('cms::simple.module')
+                ->with('form', $form->view('card'))
+                ->with('title', $this->getTitle('Просмотр'))
+                ->with('description', $this->getDescription());
+        else
+            return view('cms::module')
+                ->with('form', $form->view('card'))
+                ->with('title', $this->getTitle('Просмотр'))
+                ->with('description', $this->getDescription());
     }
 
     /**
@@ -339,7 +409,7 @@ class ModuleController extends BaseController
         }
 
         /**
-         * @var Model $this->model
+         * @var Model $this- >model
          */
         if ($this->objectId) {
             // update
@@ -366,7 +436,7 @@ class ModuleController extends BaseController
         }
 
         // Files
-        if(request()->allFiles()) {
+        if (request()->allFiles()) {
             $attachments = [];
             foreach (request()->allFiles() as $key => $files) {
                 $field = $this->getFieldByKey($key);
@@ -400,7 +470,7 @@ class ModuleController extends BaseController
                     }
                 }
 
-                $attachmentIds = array_map(function($item) {
+                $attachmentIds = array_map(function ($item) {
                     return $item->id;
                 }, $attachments);
 
@@ -429,10 +499,12 @@ class ModuleController extends BaseController
      */
     public function destroy($objectId)
     {
-        $model = $this->className::findOrFail($objectId);
-        $model->delete();
+        if ($this->canDelete) {
+            $model = $this->className::findOrFail($objectId);
+            $model->delete();
 
-        Toast::success('Запись удалена');
+            Toast::success('Запись удалена');
+        }
 
         return redirect(
             route('cms.module.index', ['controller' => $this->getSectionController()], false)
